[2025.05.12] - 김건우 1. 매도 로직 수정 (매수한 주식의 개수만큼만 매도할 수 있도록) 미완성

+++ src/main/java/com/kwfw/findiary/model/DiaryDto.java
@@ -8,7 +8,7 @@ public class DiaryDto {
     private int trading_count;
     private int status;
 
-    private float trading_price;
+    private long trading_price;
 
     private String ticker;
     private String stock_name;
